renamed the specifications and removed the old specs ones. adapted the code to the latest 1.3 (with mustEqual)
