--- src/test/scala/com/codecommit/antixml/NodeSpecs2.scala
@@ -1,99 +1,0 @@
-/*
- * Copyright (c) 2011, Daniel Spiewak
- * All rights reserved.
- * 
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- * 
- * - Redistributions of source code must retain the above copyright notice, this
- *   list of conditions and the following disclaimer. 
- * - Redistributions in binary form must reproduce the above copyright notice, this
- *   list of conditions and the following disclaimer in the documentation and/or
- *   other materials provided with the distribution.
- * - Neither the name of the <ORGANIZATION> nor the names of its contributors may
- *   be used to endorse or promote products derived from this software without
- *   specific prior written permission.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.codecommit.antixml
-
-import org.specs2.mutable._
-import org.specs2.matcher.DataTables
-
-class NodeSpecs2 extends Specification with DataTables {
-    
-  "elements" should {
-    "serialize empty elements correctly" in {
-      (<br/>).anti.toString must_== "<br/>"
-    }
-    
-    "escape reserved characters in the name" in {
-      "character" || "elem.toString" |>
-      "\""        !! "<&quot;/>"    |
-      "&"         !! "<&amp;/>"     |
-      "'"         !! "<&apos;/>"    |
-      "<"         !! "<&lt;/>"      |
-      ">"         !! "<&gt;/>"      | { (c, r) => Elem(None, c, Map(), Group()).toString must_== r }
-    }
-    
-    "escape reserved characters in the namespace" in {
-      "character" || "elem.toString" |>
-      "\""        !! "<&quot;:foo/>" |
-      "&"         !! "<&amp;:foo/>"  |
-      "'"         !! "<&apos;:foo/>" |
-      "<"         !! "<&lt;:foo/>"   |
-      ">"         !! "<&gt;:foo/>"   | { (c, r) => Elem(Some(c), "foo", Map(), Group()).toString must_== r }
-    }
-    
-    "escape reserved characters in attribute keys" in {
-      "character" || "elem.toString"         |>
-      "\""        !! "<foo &quot;=\"bar\"/>" |
-      "&"         !! "<foo &amp;=\"bar\"/>"  |
-      "'"         !! "<foo &apos;=\"bar\"/>" |
-      "<"         !! "<foo &lt;=\"bar\"/>"   |
-      ">"         !! "<foo &gt;=\"bar\"/>"   | { (c, r) => Elem(None, "foo", Map(c -> "bar"), Group()).toString must_== r }
-    }
-    
-    "escape reserved characters in attribute values" in {
-      "character" || "elem.toString"         |>
-      "\""        !! "<foo bar=\"&quot;\"/>" |
-      "&"         !! "<foo bar=\"&amp;\"/>"  |
-      "'"         !! "<foo bar=\"&apos;\"/>" |
-      "<"         !! "<foo bar=\"&lt;\"/>"   |
-      ">"         !! "<foo bar=\"&gt;\"/>"   | { (c, r) => Elem(None, "foo", Map("bar" -> c), Group()).toString must_== r }
-    }
-
-    "select against self" in {
-      val bookstore = <bookstore><book><title>For Whom the Bell Tolls</title><author>Hemmingway</author></book><book><title>I, Robot</title><author>Isaac Asimov</author></book><book><title>Programming Scala</title><author>Dean Wampler</author><author>Alex Payne</author></book></bookstore>.anti
-      (bookstore \ "book") must be_===(bookstore.children)
-      (bookstore \\ "title") must_== (bookstore.children \\ "title")
-    }
-    
-    "select text within self" in {
-      ((<parent>Text</parent>).anti \\ text mkString) must_== "Text"
-    }
-  }
-  
-  "text nodes" should {
-    "escape reserved characters when serialized" in {
-      Text("Lorem \" ipsum & dolor ' sit < amet > blargh").toString must_== "Lorem &quot; ipsum &amp; dolor &apos; sit &lt; amet &gt; blargh"
-    }
-  }
-  
-  "cdata nodes" should {
-    "not escape reserved characters when serialized" in {
-      CDATA("Lorem \" ipsum & dolor ' sit < amet > blargh").toString must_== "<![CDATA[Lorem \" ipsum & dolor ' sit < amet > blargh]]>"
-    }
-  }
-}
