+++ src/test/scala/com/codecommit/antixml/SelectorSpecs.scala
@@ -28,9 +28,10 @@
 
 package com.codecommit.antixml
 
-import org.specs._
+import org.specs2.mutable._
 
-object SelectorSpecs extends Specification {
+class SelectorSpecs extends Specification {
+  
   "the * selector" should {
     "select nothing when parent is empty" in {
       XML.fromString("<parent/>") \ * mustEqual Group()
