--- src/test/scala/com/codecommit/antixml/SelectorSpecs2.scala
@@ -1,58 +1,0 @@
-/*
- * Copyright (c) 2011, Daniel Spiewak
- * All rights reserved.
- * 
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- * 
- * - Redistributions of source code must retain the above copyright notice, this
- *   list of conditions and the following disclaimer. 
- * - Redistributions in binary form must reproduce the above copyright notice, this
- *   list of conditions and the following disclaimer in the documentation and/or
- *   other materials provided with the distribution.
- * - Neither the name of the <ORGANIZATION> nor the names of its contributors may
- *   be used to endorse or promote products derived from this software without
- *   specific prior written permission.
- * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package com.codecommit.antixml
-
-import org.specs2.mutable._
-
-class SelectorSpecs2 extends Specification {
-  
-  "the * selector" should {
-    "select nothing when parent is empty" in {
-      XML.fromString("<parent/>") \ * must be_===(Group())
-    }
-    
-    "select entire contents of parent" in {
-      val xml = XML.fromString("<parent><child1/>Test<child2/>text here we go \n with whitespace<child3>Inside!</child3></parent>")
-      val expected = xml.children
-      xml \ * must be_===(expected)
-    }
-  }
-  
-  "the element selector(s)" should {
-    "select nothing when parent is empty" in {
-      (<parent/>).anti \ "parent" must be_===(Group())
-      (<parent/>).anti \ 'parent must be_===(Group())
-    }
-    
-    "select only the named element(s)" in {
-      (<parent><foo/><bar/>Baz<foo/></parent>).anti \ "foo" must be_===(Group(<foo/>.anti, <foo/>.anti))
-      (<parent><foo/><bar/>Baz<foo/></parent>).anti \ 'foo must be_===(Group(<foo/>.anti, <foo/>.anti))
-    }
-  }
-}
