+++ src/test/scala/com/codecommit/antixml/util/VectorCaseSpecs.scala
@@ -28,22 +28,17 @@
 
 package com.codecommit.antixml.util
 
-import org.specs._
-import org.scalacheck._
+import org.specs2.mutable._
+import org.specs2.ScalaCheck
+import org.specs2.matcher.Parameters
+import org.scalacheck.{Arbitrary, Prop}
+import Prop._
+import org.specs2.matcher.ScalaCheckMatchers._
 
-object VectorCaseSpecs extends Specification with ScalaCheck {
-  import Prop._
+class VectorCaseSpecs extends Specification with ScalaCheck {
   import math._
   
-  val numProcessors = Runtime.getRuntime.availableProcessors
-  
   val vector = VectorCase[Int]()
-  
-  implicit def arbitraryVectorCase[A](implicit arb: Arbitrary[A]): Arbitrary[VectorCase[A]] = {
-    Arbitrary(for {
-      data <- Arbitrary.arbitrary[List[A]]
-    } yield data.foldLeft(VectorCase[A]()) { _ :+ _ })
-  }
   
   "VectorCase" should {
     "store a single element" in {
@@ -51,301 +46,208 @@ object VectorCaseSpecs extends Specification with ScalaCheck {
       v2(0) mustEqual 42
     }
     
-    "implement length" in {
-      val prop = forAll { (list: List[Int]) => 
-        val vec = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
-        vec.length == list.length
-      }
-      
-      prop must pass(set(workers -> numProcessors))
+    "implement length" in check { list: List[Int] =>
+      val vec = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
+      vec.length === list.length
     }
     
-    "replace single element" in {
-      val prop = forAll { (vec: VectorCase[Int], i: Int) =>
-        (!vec.isEmpty && i > Int.MinValue) ==> {
-          val idx = abs(i) % vec.length
-          val newVectorCase = vec.updated(idx, "test").updated(idx, "newTest")
-          newVectorCase(idx) == "newTest"
-        }
+    "replace single element" in check { (vec: VectorCase[Int], i: Int) =>
+      (vec.nonEmpty && i > Int.MinValue) ==> {
+        val idx = abs(i) % vec.length
+        val newVectorCase = vec.updated(idx, "test").updated(idx, "newTest")
+        newVectorCase(idx) mustEqual "newTest"
       }
-      
-      prop must pass(set(workers -> numProcessors))
+    }
+    "fail on apply out-of-bounds" in check { (vec: VectorCase[Int], i: Int) =>
+      !((0 until vec.length) contains i) ==> { vec(i) must throwA[Throwable] }
+    }
+
+    "fail on update out-of-bounds" in check { (vec: VectorCase[Int], i: Int) =>
+      !((0 until vec.length) contains i) ==> { vec.updated(i, 42) must throwA[Throwable] }
     }
     
-    "fail on apply out-of-bounds" in {
-      val prop = forAll { (vec: VectorCase[Int], i: Int) =>
-        !((0 until vec.length) contains i) ==> {
-          vec(i) must throwA[Throwable]
-        }
-      }
-      
-      prop must pass
+    "store multiple elements in order" in check { list: List[Int] =>
+      val newVectorCase = list.foldLeft(vector) { _ :+ _ }
+      val res = for (i <- 0 until list.length) yield newVectorCase(i) == list(i)
+
+      res must not contain (false)
     }
-    
-    "fail on update out-of-bounds" in {
-      val prop = forAll { (vec: VectorCase[Int], i: Int) =>
-        !((0 to vec.length) contains i) ==> {
-          vec.updated(i, 42) must throwA[Throwable]
-        }
-      }
-      
-      prop must pass(set(workers -> numProcessors))
-    }
-    
-    "store multiple elements in order" in {
-      val prop = forAll { list: List[Int] =>
-        val newVectorCase = list.foldLeft(vector) { _ :+ _ }
-        val res = for (i <- 0 until list.length) yield newVectorCase(i) == list(i)
-        
-        res forall { _ == true }
-      }
-      
-      prop must pass(set(workers -> numProcessors))
-    }
-    
+
     "store lots of elements" in {
       val LENGTH = 100000
       val vector = (0 until LENGTH).foldLeft(VectorCase[Int]()) { _ :+ _ }
       
       vector.length mustEqual LENGTH
-      for (i <- 0 until LENGTH) {
-        vector(i) mustEqual i
-      }
+      ((i:Int) => vector(i) mustEqual i).forall(0 until LENGTH)
     }
     
-    "maintain both old and new versions after conj" in {
-      val prop = forAll { vec: VectorCase[Int] =>
-        val vec2 = vec :+ 42
-        for (i <- 0 until vec.length) {
+    "maintain both old and new versions after conj" in check { vec: VectorCase[Int] =>
+      val vec2 = vec :+ 42
+      for (i <- 0 until vec.length) {
+        vec2(i) aka ("Index " + i + " in derivative") mustEqual vec(i) aka ("Index " + i + " in origin")
+      }
+      vec2.last mustEqual 42
+    }.set(maxSize -> 3000, minTestsOk -> 1000, workers -> numProcessors)
+
+    "maintain both old and new versions after update" in check { (vec: VectorCase[Int], i: Int) =>
+      (!vec.isEmpty && i > Int.MinValue) ==> {
+        val idx = abs(i) % vec.length
+        val vec2 = vec.updated(idx, 42)
+        for (i <- 0 until vec.length if i != idx) {
           vec2(i) aka ("Index " + i + " in derivative") mustEqual vec(i) aka ("Index " + i + " in origin")
         }
-        vec2.last mustEqual 42
+        vec2(idx) mustEqual 42
       }
-      
-      prop must pass(set(maxSize -> 3000, minTestsOk -> 1000, workers -> numProcessors))
     }
-    
-    "maintain both old and new versions after update" in {
-      val prop = forAll { (vec: VectorCase[Int], i: Int) =>
-        (!vec.isEmpty && i > Int.MinValue) ==> {
-          val idx = abs(i) % vec.length
-          val vec2 = vec.updated(idx, 42)
-          for (i <- 0 until vec.length if i != idx) {
-            vec2(i) aka ("Index " + i + " in derivative") mustEqual vec(i) aka ("Index " + i + " in origin")
-          }
-          vec2(idx) mustEqual 42
+
+    "implement filter" in check { (vec: VectorCase[Int], f: (Int)=>Boolean) =>
+      val filtered = vec filter f
+
+      var back = filtered forall f
+      for (e <- vec) {
+        if (f(e)) {
+          back &&= filtered.contains(e)
         }
       }
-      
-      prop must pass(set(maxSize -> 3000, minTestsOk -> 1000, workers -> numProcessors))
+      back
     }
     
-    "implement filter" in {
-      val prop = forAll { (vec: VectorCase[Int], f: (Int)=>Boolean) =>
-        val filtered = vec filter f
-        
-        var back = filtered forall f
-        for (e <- vec) {
-          if (f(e)) {
-            back &&= filtered.contains(e)
-          }
-        }
-        back
+    "implement foldLeft" in check { list: List[Int] =>
+      val vec = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
+      vec.foldLeft(0) { _ + _ } === list.foldLeft(0) { _ + _ }
+    }
+    
+    "implement forall" in check { (vec: VectorCase[Int], f: (Int)=>Boolean) =>
+      val bool = vec forall f
+      
+      var back = true
+      for (e <- vec) {
+        back &&= f(e)
       }
       
-      prop must pass(set(workers -> numProcessors))
+      (back && bool) || (!back && !bool)
     }
     
-    "implement foldLeft" in {
-      val prop = forAll { list: List[Int] =>
-        val vec = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
-        vec.foldLeft(0) { _ + _ } == list.foldLeft(0) { _ + _ }
-      }
+    "implement flatMap" in check { (vec: VectorCase[Int], f: (Int)=>VectorCase[Int]) =>
+      val mapped = vec flatMap f
       
-      prop must pass(set(workers -> numProcessors))
-    }
-    
-    "implement forall" in {
-      val prop = forAll { (vec: VectorCase[Int], f: (Int)=>Boolean) =>
-        val bool = vec forall f
-        
-        var back = true
-        for (e <- vec) {
-          back &&= f(e)
-        }
-        
-        (back && bool) || (!back && !bool)
-      }
+      var back = true
       
-      prop must pass(set(workers -> numProcessors))
-    }
-    
-    "implement flatMap" in {
-      val prop = forAll { (vec: VectorCase[Int], f: (Int)=>VectorCase[Int]) =>
-        val mapped = vec flatMap f
+      var i = 0
+      var n = 0
+      
+      while (i < vec.length) {
+        val res = f(vec(i))
         
-        var back = true
-        
-        var i = 0
-        var n = 0
-        
-        while (i < vec.length) {
-          val res = f(vec(i))
+        var inner = 0
+        while (inner < res.length) {
+          back &&= mapped(n) == res(inner)
           
-          var inner = 0
-          while (inner < res.length) {
-            back &&= mapped(n) == res(inner)
-            
-            inner += 1
-            n += 1
-          }
-          
-          i += 1
+          inner += 1
+          n += 1
         }
         
-        back
+        i += 1
       }
-      
-      prop must pass(set(workers -> numProcessors))
+
+      back
     }
     
-    "implement map" in {
-      val prop = forAll { (vec: VectorCase[Int], f: (Int)=>Int) =>
-        val mapped = vec map f
-        
-        var back = vec.length == mapped.length
-        for (i <- 0 until vec.length) {
-          back &&= mapped(i) == f(vec(i))
-        }
-        back
-      }
+    "implement map" in check { (vec: VectorCase[Int], f: (Int)=>Int) =>
+      val mapped = vec map f
       
-      prop must pass(set(workers -> numProcessors))
+      var back = vec.length == mapped.length
+      for (i <- 0 until vec.length) {
+        back &&= mapped(i) == f(vec(i))
+      }
+      back
     }
     
-    "implement reverse" in {
-      val prop = forAll { v: VectorCase[Int] =>
-        val reversed = v.reverse
-        
-        var back = v.length == reversed.length
-        for (i <- 0 until v.length) {
-          back &&= reversed(i) == v(v.length - i - 1)
-        }
-        back
-      }
+    "implement reverse" in check { v: VectorCase[Int] =>
+      val reversed = v.reverse
       
-      prop must pass(set(workers -> numProcessors))
+      var back = v.length == reversed.length
+      for (i <- 0 until v.length) {
+        back &&= reversed(i) == v(v.length - i - 1)
+      }
+      back
     }
     
-    "append to reverse" in {
-      val prop = forAll { (v: VectorCase[Int], n: Int) =>
-        val rev = v.reverse
-        val add = rev :+ n
-        
-        var back = add.length == rev.length + 1
-        for (i <- 0 until rev.length) {
-          back &&= add(i) == rev(i)
-        }
-        back && add(rev.length) == n
-      }
+    "append to reverse" in check { (v: VectorCase[Int], n: Int) =>
+      val rev = v.reverse
+      val add = rev :+ n
       
-      prop must pass(set(workers -> numProcessors))
+      var back = add.length == rev.length + 1
+      for (i <- 0 until rev.length) {
+        back &&= add(i) == rev(i)
+      }
+      back && add(rev.length) === n
     }
     
-    "map on reverse" in {
-      val prop = forAll { (v: VectorCase[Int], f: (Int)=>Int) =>
-        val rev = v.reverse
-        val mapped = rev map f
-        
-        var back = mapped.length == rev.length
-        for (i <- 0 until rev.length) {
-          back &&= mapped(i) == f(rev(i))
-        }
-        back
-      }
+    "map on reverse" in check { (v: VectorCase[Int], f: (Int)=>Int) =>
+      val rev = v.reverse
+      val mapped = rev map f
       
-      prop must pass(set(workers -> numProcessors))
+      var back = mapped.length == rev.length
+      for (i <- 0 until rev.length) {
+        back &&= mapped(i) == f(rev(i))
+      }
+      back
     }
     
-    "implement zip" in {
-      val prop = forAll { (first: VectorCase[Int], second: VectorCase[Double]) =>
-        val zip = first zip second
-        
-        var back = zip.length == min(first.length, second.length)
-        for (i <- 0 until zip.length) {
-          var (left, right) = zip(i)
-          back &&= (left == first(i) && right == second(i))
-        }
-        back
-      }
+    "implement zip" in check { (first: VectorCase[Int], second: VectorCase[Double]) =>
+      val zip = first zip second
       
-      prop must pass(set(workers -> numProcessors))
+      var back = zip.length == min(first.length, second.length)
+      for (i <- 0 until zip.length) {
+        var (left, right) = zip(i)
+        back &&= (left == first(i) && right == second(i))
+      }
+      back
     }
     
-    "implement zipWithIndex" in {
-      val prop = forAll { vec: VectorCase[Int] =>
-        val zip = vec.zipWithIndex
-        
-        var back = zip.length == vec.length
-        for (i <- 0 until zip.length) {
-          val (elem, index) = zip(i)
-          
-          back &&= (index == i && elem == vec(i))
-        }
-        back
-      }
+    "implement zipWithIndex" in check { vec: VectorCase[Int] =>
+      val zip = vec.zipWithIndex
       
-      prop must pass(set(workers -> numProcessors))
+      var back = zip.length == vec.length
+      for (i <- 0 until zip.length) {
+        val (elem, index) = zip(i)
+        back &&= (index == i && elem == vec(i))
+      }
+      back
     }
     
-    "implement equals" in {
-      {
-        val prop = forAll { list: List[Int] => 
-          val vecA = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
-          val vecB = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
-          
-          vecA == vecB
-        }
-        
-        prop must pass(set(workers -> numProcessors))
-      }
-      
-      {
-        val prop = forAll { (vecA: VectorCase[Int], vecB: VectorCase[Int]) =>
-          vecA.length != vecB.length ==> (vecA != vecB)
-        }
-        
-        prop must pass(set(workers -> numProcessors))
-      }
-      
-      {
-        val prop = forAll { (listA: List[Int], listB: List[Int]) =>
-          val vecA = listA.foldLeft(VectorCase[Int]()) { _ :+ _ }
-          val vecB = listB.foldLeft(VectorCase[Int]()) { _ :+ _ }
-          
-          listA != listB ==> (vecA != vecB)
-        }
-        
-        prop must pass(set(workers -> numProcessors))
-      }
-      
-      {
-        val prop = forAll { (vec: VectorCase[Int], data: Int) => vec != data }
-        
-        prop must pass(set(workers -> numProcessors))
-      }
-    }
-    
-    "implement hashCode" in {
-      val prop = forAll { list: List[Int] =>
+    "implement equals" >> {
+      "1." in check { list: List[Int] => 
         val vecA = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
         val vecB = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
-        
-        vecA.hashCode == vecB.hashCode
+        vecA === vecB
       }
-      
-      prop must pass(set(workers -> numProcessors))
+      "2." in check { (vecA: VectorCase[Int], vecB: VectorCase[Int]) =>
+        vecA.length != vecB.length ==> (vecA != vecB)
+      }
+      "3." in check { (listA: List[Int], listB: List[Int]) =>
+        val vecA = listA.foldLeft(VectorCase[Int]()) { _ :+ _ }
+        val vecB = listB.foldLeft(VectorCase[Int]()) { _ :+ _ }
+        
+        listA != listB ==> (vecA != vecB)
+      }      
+      "4." in check { (vec: VectorCase[Int], data: Int) => vec !== data }
+    }
+    
+    "implement hashCode" in check { list: List[Int] =>
+      val vecA = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
+      val vecB = list.foldLeft(VectorCase[Int]()) { _ :+ _ }
+      vecA.hashCode === vecB.hashCode
     }
   }
+  
+  implicit def arbitraryVectorCase[A](implicit arb: Arbitrary[A]): Arbitrary[VectorCase[A]] = {
+    Arbitrary(for {
+      data <- Arbitrary.arbitrary[List[A]]
+    } yield data.foldLeft(VectorCase[A]()) { _ :+ _ })
+  }
+  val numProcessors = Runtime.getRuntime.availableProcessors
 }
 
+
